search_indices: move RecentDocsActionMessage to file
