+++ x-pack/plugins/search_indices/public/components/index_documents/index_documents.tsx
@@ -9,68 +9,17 @@ import React from 'react';
 
 import { Result, resultToField, resultMetaData } from '@kbn/search-index-documents';
 
-import { i18n } from '@kbn/i18n';
-import {
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiIcon,
-  EuiLink,
-  EuiPanel,
-  EuiProgress,
-  EuiSpacer,
-} from '@elastic/eui';
+import { EuiFlexGroup, EuiFlexItem, EuiPanel, EuiProgress, EuiSpacer } from '@elastic/eui';
 import { useIndexDocumentSearch } from '../../hooks/api/use_document_search';
 import { useIndexMapping } from '../../hooks/api/use_index_mappings';
-import { useKibana } from '../../hooks/use_kibana';
 import { AddDocumentsCodeExample } from './add_documents_code_example';
+
+import { DEFAULT_PAGE_SIZE } from './constants';
+import { RecentDocsActionMessage } from './recent_docs_action_message';
 
 interface IndexDocumentsProps {
   indexName: string;
 }
-
-interface RecentDocsActionMessageProps {
-  indexName: string;
-}
-
-const DEFAULT_PAGE_SIZE = 50;
-
-const RecentDocsActionMessage: React.FC<RecentDocsActionMessageProps> = ({ indexName }) => {
-  const {
-    services: { share },
-  } = useKibana();
-
-  const discoverLocator = share.url.locators.get('DISCOVER_APP_LOCATOR');
-
-  const onClick = async () => {
-    await discoverLocator?.navigate({ dataViewSpec: { title: indexName } });
-  };
-
-  return (
-    <EuiPanel hasBorder={false} hasShadow={false} color="subdued" borderRadius="none">
-      <EuiFlexGroup>
-        <EuiFlexItem grow={false}>
-          <EuiIcon type="calendar" />
-        </EuiFlexItem>
-        <EuiFlexItem>
-          <p>
-            {i18n.translate('xpack.searchIndices.indexDocuments.recentDocsActionMessage', {
-              defaultMessage:
-                'You are viewing the {pageSize} most recently ingested documents in this index. To see all documents, view in',
-              values: {
-                pageSize: DEFAULT_PAGE_SIZE,
-              },
-            })}{' '}
-            <EuiLink onClick={onClick}>
-              {i18n.translate('xpack.searchIndices.indexDocuments.recentDocsActionMessageLink', {
-                defaultMessage: 'Discover.',
-              })}
-            </EuiLink>
-          </p>
-        </EuiFlexItem>
-      </EuiFlexGroup>
-    </EuiPanel>
-  );
-};
 
 export const IndexDocuments: React.FC<IndexDocumentsProps> = ({ indexName }) => {
   const { data: indexDocuments, isInitialLoading } = useIndexDocumentSearch(indexName, {
@@ -89,7 +38,7 @@ export const IndexDocuments: React.FC<IndexDocumentsProps> = ({ indexName }) => 
       <EuiFlexGroup direction="column">
         <EuiFlexItem>
           {isInitialLoading && <EuiProgress size="xs" color="primary" />}
-          {docs.length === 0 && <AddDocumentsCodeExample />}
+          {!isInitialLoading && docs.length === 0 && <AddDocumentsCodeExample />}
           {docs.length > 0 && (
             <>
               <RecentDocsActionMessage indexName={indexName} />
